requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
pgvector==0.2.0
psycopg2-binary==2.9.6
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
